Make bump mapping a cvar, disable ssao by default, remove test2 from menu

+++ src/graphics/GraphicsManager.cc
@@ -130,7 +130,7 @@ namespace sp
 			defaultView.extents = CVarWindowSize.Get();
 			views.push_back(defaultView);
 		}
-		for (uint i = 1; i < playerViews.size(); i++)
+		for (size_t i = 1; i < playerViews.size(); i++)
 		{
 			views.push_back(*ecs::UpdateViewCache(playerViews[i]));
 		}
